# server/src/main/resources/openapi/family_accounting_openapi.yaml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  description: |
    家庭记账系统的 RESTful API，支持用户认证、人员管理、账户管理、交易记录和统计。
    支持银行、微信、支付宝等多种账号，每种账号下可有多张卡或子账户。
    角色权限：
    - 管理员：访问所有数据，管理用户、账号、子账户。
    - 普通用户：仅访问通过 user_sub_accounts 关联的子账户和自己的交易/统计。
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
...
paths:
...
  '/transactions/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Transactions
        - transactions
      summary: 获取交易记录详情
      operationId: getTransaction
      parameters: []
      responses:
        '200':
          description: 交易记录详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    put:
      tags:
        - Transactions
        - transactions
      summary: 更新交易记录
      operationId: updateTransaction
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
        required: true
      responses:
        '200':
          description: 交易记录已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '400':
          description: 无效输入（交易类型不存在或参数格式错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Transactions
        - transactions
      summary: 删除交易记录
      operationId: deleteTransaction
      parameters: []
      responses:
        '204':
          description: 交易记录已删除
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
...
  '/sub-accounts/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - sub-accounts
        - SubAccounts
      summary: 获取子账户详情
      operationId: getSubAccount
      parameters: []
      responses:
        '200':
          description: 子账户详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubAccountDTO'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 子账户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    put:
      tags:
        - sub-accounts
        - SubAccounts
      summary: 更新子账户
      operationId: subAccountsUpdate
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubAccountRequest'
        required: true
      responses:
        '200':
          description: 子账户已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubAccount'
        '400':
          description: 无效输入（账户编号重复或参数格式错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 子账户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    delete:
      tags:
        - sub-accounts
        - SubAccounts
      summary: 删除子账户
      operationId: subAccountsDelete
      parameters: []
      responses:
        '204':
          description: 子账户已删除
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 子账户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 子账户存在关联交易，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
...
  /transactions:
    get:
      tags:
        - Transactions
        - transactions
      summary: 列出交易记录
      operationId: transactionsList
      parameters:
        - name: page
          in: query
          required: true
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: size
          in: query
          required: true
          schema:
            maximum: 100
            minimum: 1
            type: integer
            format: int32
        - name: userId
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: subAccountId
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 交易记录列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    post:
      tags:
        - Transactions
        - transactions
      summary: 创建交易记录
      operationId: transactionsCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
        required: true
      responses:
        '201':
          description: 交易记录已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '400':
          description: 无效输入（交易类型不存在或参数格式错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（普通用户尝试访问未关联的子账户）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    ApiError:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
          enum:
            - PERMISSION_DENIED
            - DUPLICATE_ACCOUNT
            - INVALID_TRANSACTION
            - DATA_NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
            - INVALID_REQUEST
            - VALIDATION_FAILED
            - INVALID_CREDENTIALS
        message:
          type: string
...
    TransactionDTO:
      required:
        - amount
        - subAccountId
        - time
        - transactionId
        - typeId
        - userId
      type: object
      properties:
        transactionId:
          type: integer
          format: int32
        time:
          type: string
          format: date-time
        typeId:
          type: integer
          format: int32
        isIncome:
          type: boolean
          description: 是否为收入（由typeId关联的transaction_types.is_income自动生成）
        subAccountId:
          type: integer
          format: int32
        amount:
          pattern: '^\d+(\.\d{2})?$'
          type: string
        userId:
          type: integer
          format: int32
        remarks:
          maxLength: 255
          minLength: 0
          type: string
    TransactionRequest:
      required:
        - amount
        - subAccountId
        - time
        - typeId
      type: object
      properties:
        time:
          type: string
          format: date-time
        typeId:
          type: integer
          format: int32
        subAccountId:
          type: integer
          format: int32
        amount:
          pattern: '^\d+(\.\d{2})?$'
          type: string
        remarks:
          maxLength: 255
          minLength: 0
          type: string
...
    Account:
      type: object
      properties:
        accountId:
          type: integer
          format: int32
        accountType:
          $ref: '#/components/schemas/AccountType'
        accountName:
          type: string
        subAccounts:
          type: array
          items:
            $ref: '#/components/schemas/SubAccount'
    AccountType:
      type: object
      properties:
        typeId:
          type: integer
          format: int32
        typeName:
          type: string
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
    SubAccount:
      type: object
      properties:
        subAccountId:
          type: integer
          format: int32
        account:
          $ref: '#/components/schemas/Account'
        accountName:
          type: string
        accountNumber:
          type: string
        cardType:
          type: string
          enum:
            - SAVINGS
            - DEBIT
            - CREDIT
            - VISA
            - MASTERCARD
            - WALLET
        balance:
          type: number
        users:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/User'
    Transaction:
      type: object
      properties:
        transactionId:
          type: integer
          format: int32
        time:
          type: string
          format: date-time
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        subAccount:
          $ref: '#/components/schemas/SubAccount'
        amount:
          type: number
        user:
          $ref: '#/components/schemas/User'
        remarks:
          type: string
    TransactionType:
      type: object
      properties:
        typeId:
          type: integer
          format: int32
        typeName:
          type: string
        isIncome:
          type: boolean
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        username:
          type: string
        password:
          type: string
        isAdmin:
          type: boolean
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    SubAccountRequest:
      required:
        - accountId
        - accountName
        - accountNumber
        - balance
        - cardType
      type: object
      properties:
        accountId:
          type: integer
          format: int32
        accountName:
          maxLength: 100
          minLength: 0
          type: string
        accountNumber:
          maxLength: 50
          minLength: 0
          type: string
        cardType:
          type: string
          enum:
            - SAVINGS
            - DEBIT
            - CREDIT
            - VISA
            - MASTERCARD
            - WALLET
        balance:
          pattern: '^\d+(\.\d{2})?$'
          type: string
...
    TransactionListResponse:
      required:
        - total
        - transactions
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDTO'
        total:
          type: integer
          format: int32
    SubAccountDTO:
      required:
        - accountId
        - accountName
        - accountNumber
        - balance
        - cardType
        - subAccountId
      type: object
      properties:
        subAccountId:
          type: integer
          format: int32
        accountId:
          type: integer
          format: int32
        accountName:
          maxLength: 100
          minLength: 0
          type: string
        accountNumber:
          maxLength: 50
          minLength: 0
          type: string
        cardType:
          type: string
          enum:
            - SAVINGS
            - DEBIT
            - CREDIT
            - VISA
            - MASTERCARD
            - WALLET
        balance:
          pattern: '^\d+(\.\d{2})?$'
          type: string
...